+++ src/main/kotlin/com/alex/eyk/telegram/data/entity/competition/EducationBasis.kt
@@ -1,4 +1,4 @@
-package com.alex.eyk.telegram.model.entity.competition
+package com.alex.eyk.telegram.data.entity.competition
 
 enum class EducationBasis(
     val code: Int
