+++ src/main/kotlin/com/alex/eyk/telegram/data/entity/recent/RecentDirectionRepository.kt
@@ -1,4 +1,4 @@
-package com.alex.eyk.telegram.model.entity.recent
+package com.alex.eyk.telegram.data.entity.recent
 
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
