+++ src/main/kotlin/com/alex/eyk/telegram/data/entity/user/UserRepository.kt
@@ -1,4 +1,4 @@
-package com.alex.eyk.telegram.model.entity.user
+package com.alex.eyk.telegram.data.entity.user
 
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
