+++ src/main/kotlin/com/alex/eyk/telegram/data/validation/ValidationException.kt
@@ -1,3 +1,3 @@
-package com.alex.eyk.telegram.model.validation
+package com.alex.eyk.telegram.data.validation
 
 abstract class ValidationException : Exception()
