+++ src/main/kotlin/com/alex/eyk/telegram/data/validation/impl/WrongFormatException.kt
@@ -1,0 +1,5 @@
+package com.alex.eyk.telegram.data.validation.impl
+
+import com.alex.eyk.telegram.data.validation.ValidationException
+
+class WrongFormatException : ValidationException()
