+++ src/main/kotlin/com/alex/eyk/telegram/handler/analytics/AbstractAnalyticsHandler.kt
@@ -4,16 +4,16 @@ import com.alex.eyk.dictionary.builder.AnalyzeResultArgumentsBuilder
 import com.alex.eyk.dictionary.keys.Replies
 import com.alex.eyk.dictionary.keys.Words
 import com.alex.eyk.replies.dictionary.provider.DictionaryProvider
-import com.alex.eyk.telegram.model.entity.competition.Competition
-import com.alex.eyk.telegram.model.entity.competition.Direction
-import com.alex.eyk.telegram.model.entity.competition.EducationBasis
-import com.alex.eyk.telegram.model.entity.recent.RecentDirection
-import com.alex.eyk.telegram.model.entity.recent.RecentDirectionRepository
-import com.alex.eyk.telegram.model.entity.user.Activity
-import com.alex.eyk.telegram.model.entity.user.User
-import com.alex.eyk.telegram.model.entity.user.UserRepository
-import com.alex.eyk.telegram.model.validation.Result
-import com.alex.eyk.telegram.model.validation.impl.DirectionCodeValidator
+import com.alex.eyk.telegram.data.entity.competition.Competition
+import com.alex.eyk.telegram.data.entity.competition.Direction
+import com.alex.eyk.telegram.data.entity.competition.EducationBasis
+import com.alex.eyk.telegram.data.entity.recent.RecentDirection
+import com.alex.eyk.telegram.data.entity.recent.RecentDirectionRepository
+import com.alex.eyk.telegram.data.entity.user.Activity
+import com.alex.eyk.telegram.data.entity.user.User
+import com.alex.eyk.telegram.data.entity.user.UserRepository
+import com.alex.eyk.telegram.data.validation.Result
+import com.alex.eyk.telegram.data.validation.impl.DirectionCodeValidator
 import com.alex.eyk.telegram.service.analyze.AnaliticsService
 import com.alex.eyk.telegram.service.analyze.CompetitionAnalytics
 import com.alex.eyk.telegram.service.analyze.exception.ParticipantNotFoundException
