+++ src/main/kotlin/com/alex/eyk/telegram/handler/analytics/PaidAnalyticsHandler.kt
@@ -1,10 +1,10 @@
 package com.alex.eyk.telegram.handler.analytics
 
 import com.alex.eyk.replies.dictionary.provider.DictionaryProvider
-import com.alex.eyk.telegram.model.entity.competition.EducationBasis
-import com.alex.eyk.telegram.model.entity.recent.RecentDirectionRepository
-import com.alex.eyk.telegram.model.entity.user.Activity
-import com.alex.eyk.telegram.model.entity.user.UserRepository
+import com.alex.eyk.telegram.data.entity.competition.EducationBasis
+import com.alex.eyk.telegram.data.entity.recent.RecentDirectionRepository
+import com.alex.eyk.telegram.data.entity.user.Activity
+import com.alex.eyk.telegram.data.entity.user.UserRepository
 import com.alex.eyk.telegram.service.analyze.AnaliticsService
 import com.alex.eyk.telegram.service.holder.CompetitionsHolder
 import org.springframework.beans.factory.annotation.Autowired
