+++ src/main/kotlin/com/alex/eyk/telegram/handler/trigger/AbstractAnaliticsTrigger.kt
@@ -2,9 +2,9 @@ package com.alex.eyk.telegram.handler.trigger
 
 import com.alex.eyk.dictionary.keys.Replies
 import com.alex.eyk.replies.dictionary.provider.DictionaryProvider
-import com.alex.eyk.telegram.model.entity.user.Activity
-import com.alex.eyk.telegram.model.entity.user.User
-import com.alex.eyk.telegram.model.entity.user.UserRepository
+import com.alex.eyk.telegram.data.entity.user.Activity
+import com.alex.eyk.telegram.data.entity.user.User
+import com.alex.eyk.telegram.data.entity.user.UserRepository
 import com.alex.eyk.telegram.util.RecentDirectionUtils
 import com.alex.eyk.telegram.util.SendMessageUtils
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage
