+++ src/main/kotlin/com/alex/eyk/telegram/handler/trigger/PaidTrigger.kt
@@ -1,8 +1,8 @@
 package com.alex.eyk.telegram.handler.trigger
 
 import com.alex.eyk.replies.dictionary.provider.DictionaryProvider
-import com.alex.eyk.telegram.model.entity.user.Activity
-import com.alex.eyk.telegram.model.entity.user.UserRepository
+import com.alex.eyk.telegram.data.entity.user.Activity
+import com.alex.eyk.telegram.data.entity.user.UserRepository
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
