--- src/main/kotlin/com/alex/eyk/telegram/model/validation/impl/WrongFormatException.kt
@@ -1,5 +1,0 @@
-package com.alex.eyk.telegram.model.validation.impl
-
-import com.alex.eyk.telegram.model.validation.ValidationException
-
-class WrongFormatException : ValidationException()
