+++ src/main/kotlin/com/alex/eyk/telegram/service/holder/CompetitionsHolder.kt
@@ -1,7 +1,7 @@
 package com.alex.eyk.telegram.service.holder
 
-import com.alex.eyk.telegram.model.entity.competition.Competition
-import com.alex.eyk.telegram.model.entity.competition.Direction
+import com.alex.eyk.telegram.data.entity.competition.Competition
+import com.alex.eyk.telegram.data.entity.competition.Direction
 
 interface CompetitionsHolder {
