+++ src/main/kotlin/com/alex/eyk/telegram/service/holder/CompetitionsHolderImpl.kt
@@ -1,8 +1,8 @@
 package com.alex.eyk.telegram.service.holder
 
 import com.alex.eyk.telegram.config.WebProperties
-import com.alex.eyk.telegram.model.entity.competition.Competition
-import com.alex.eyk.telegram.model.entity.competition.Direction
+import com.alex.eyk.telegram.data.entity.competition.Competition
+import com.alex.eyk.telegram.data.entity.competition.Direction
 import com.alex.eyk.telegram.service.loader.CompetitionsLoader
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
