+++ src/main/kotlin/com/alex/eyk/telegram/service/loader/CompetitionsLoaderImpl.kt
@@ -1,8 +1,8 @@
 package com.alex.eyk.telegram.service.loader
 
 import com.alex.eyk.telegram.config.WebProperties
-import com.alex.eyk.telegram.model.entity.competition.Competition
-import com.alex.eyk.telegram.model.entity.competition.Direction
+import com.alex.eyk.telegram.data.entity.competition.Competition
+import com.alex.eyk.telegram.data.entity.competition.Direction
 import com.alex.eyk.telegram.service.loader.parser.RawTextToParticipantsParser
 import com.alex.eyk.telegram.service.pdf.PdfService
 import com.alex.eyk.telegram.util.TextUtils
