+++ src/main/kotlin/com/alex/eyk/telegram/service/loader/parser/ParticipantsData.kt
@@ -1,7 +1,7 @@
 package com.alex.eyk.telegram.service.loader.parser
 
 import com.alex.eyk.telegram.collection.SkipList
-import com.alex.eyk.telegram.model.entity.competition.Participant
+import com.alex.eyk.telegram.data.entity.competition.Participant
 
 data class ParticipantsData(
     val participant: SkipList<Participant>,
