+++ src/main/kotlin/com/alex/eyk/telegram/telegram/handler/AbstractHandler.kt
@@ -1,7 +1,7 @@
 package com.alex.eyk.telegram.telegram.handler
 
-import com.alex.eyk.telegram.model.entity.user.Role
-import com.alex.eyk.telegram.model.entity.user.User
+import com.alex.eyk.telegram.data.entity.user.Role
+import com.alex.eyk.telegram.data.entity.user.User
 import com.alex.eyk.telegram.telegram.handler.perm.IgnoreStrategy
 import com.alex.eyk.telegram.telegram.handler.perm.PermissionDeniedStrategy
 import org.telegram.telegrambots.meta.api.methods.BotApiMethod
