+++ src/main/kotlin/com/alex/eyk/telegram/telegram/handler/HandlerProvider.kt
@@ -1,8 +1,8 @@
 package com.alex.eyk.telegram.telegram.handler
 
 import com.alex.eyk.telegram.handler.StartHandler
-import com.alex.eyk.telegram.model.entity.user.Activity
-import com.alex.eyk.telegram.model.entity.user.User
+import com.alex.eyk.telegram.data.entity.user.Activity
+import com.alex.eyk.telegram.data.entity.user.User
 import com.alex.eyk.telegram.telegram.handler.command.CommandHandlerProvider
 import com.alex.eyk.telegram.telegram.handler.message.activity.ActivityMessageHandlerProvider
 import com.alex.eyk.telegram.telegram.handler.message.condition.ConditionMessageHandlerProvider
