+++ src/main/kotlin/com/alex/eyk/telegram/telegram/handler/message/activity/ActivityMessageHandler.kt
@@ -1,6 +1,6 @@
 package com.alex.eyk.telegram.telegram.handler.message.activity
 
-import com.alex.eyk.telegram.model.entity.user.Activity
+import com.alex.eyk.telegram.data.entity.user.Activity
 import com.alex.eyk.telegram.telegram.handler.AbstractHandler
 
 abstract class ActivityMessageHandler(
