+++ src/main/kotlin/com/alex/eyk/telegram/telegram/handler/message/condition/ConditionMessageHandler.kt
@@ -1,6 +1,6 @@
 package com.alex.eyk.telegram.telegram.handler.message.condition
 
-import com.alex.eyk.telegram.model.entity.user.User
+import com.alex.eyk.telegram.data.entity.user.User
 import com.alex.eyk.telegram.telegram.handler.AbstractHandler
 import org.telegram.telegrambots.meta.api.objects.Message
