+++ src/main/kotlin/com/alex/eyk/telegram/telegram/handler/message/condition/ConditionMessageHandlerProvider.kt
@@ -1,6 +1,6 @@
 package com.alex.eyk.telegram.telegram.handler.message.condition
 
-import com.alex.eyk.telegram.model.entity.user.User
+import com.alex.eyk.telegram.data.entity.user.User
 import org.springframework.stereotype.Service
 import org.telegram.telegrambots.meta.api.objects.Message
