+++ src/main/kotlin/com/alex/eyk/telegram/telegram/handler/perm/IgnoreStrategy.kt
@@ -1,6 +1,6 @@
 package com.alex.eyk.telegram.telegram.handler.perm
 
-import com.alex.eyk.telegram.model.entity.user.User
+import com.alex.eyk.telegram.data.entity.user.User
 import com.alex.eyk.telegram.telegram.method.AbsentApiMethod
 
 class IgnoreStrategy : PermissionDeniedStrategy {
