+++ src/main/kotlin/com/alex/eyk/telegram/util/SendMessageUtils.kt
@@ -1,7 +1,7 @@
 package com.alex.eyk.telegram.util
 
 import com.alex.eyk.replies.dictionary.Reply
-import com.alex.eyk.telegram.model.entity.user.User
+import com.alex.eyk.telegram.data.entity.user.User
 import com.alex.eyk.telegram.telegram.method.SendMessageBuilder
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.ReplyKeyboardRemove
